+++ app/Traits/Operasional/SchemaMainOperasionalTableColumnsKt.php
@@ -43,16 +43,16 @@ trait SchemaMainOperasionalTableColumnsKt
       $table->integer('kode_hs')->nullable();
       $table->string('nama_komoditas')->index()->nullable();
       $table->string('nama_komoditas_en')->nullable();
-      $table->integer('volume_netto')->default(0)->index()->nullable();
+      $table->float('volume_netto')->default(0)->index()->nullable();
       $table->string('sat_netto', 50)->index()->nullable();
-      $table->integer('volume_bruto')->default(0)->nullable();
+      $table->float('volume_bruto')->default(0)->nullable();
       $table->string('sat_bruto', 50)->nullable();
-      $table->integer('volume_lain')->default(0)->nullable();
+      $table->float('volume_lain')->default(0)->nullable();
       $table->string('sat_lain', 50)->nullable();
-      $table->integer('volumeP1')->default(0)->nullable();
-      $table->integer('nettoP1')->default(0)->nullable();
-      $table->integer('volumeP8')->default(0)->nullable();
-      $table->integer('nettoP8')->default(0)->nullable();
+      $table->float('volumeP1')->default(0)->nullable();
+      $table->float('nettoP1')->default(0)->nullable();
+      $table->float('volumeP8')->default(0)->nullable();
+      $table->float('nettoP8')->default(0)->nullable();
       $table->string('dok_pelepasan', 50)->index()->nullable();
       $table->string('nomor_dok_pelepasan', 50)->index()->nullable();
       $table->date('tanggal_pelepasan')->index()->nullable();
@@ -60,8 +60,8 @@ trait SchemaMainOperasionalTableColumnsKt
       $table->string('no_kwitansi', 100)->nullable();
       $table->text('dokumen_pendukung')->nullable();
       $table->text('kontainer')->nullable();
-      $table->integer('biaya_perjadin')->default(0)->nullable();
-      $table->integer('total_pnbp')->default(0)->index()->nullable();
+      $table->float('biaya_perjadin')->default(0)->nullable();
+      $table->float('total_pnbp')->default(0)->index()->nullable();
       $table->timestamps();
 
       return $table;
